test: added the tests to the test branch
